+++ ikea-adam-service/index.js
@@ -1,0 +1,15 @@
+require("dotenv").config();
+const express = require("express");
+const cors = require("cors");
+const routes = require("./routes");
+
+const app = express();
+
+app.use(cors());
+app.use(express.json());
+app.use("/api", routes);
+
+const PORT = process.env.PORT || 3002;
+app.listen(PORT, () => {
+  console.log("\n IKEA Adam Service Started");
+});