+++ ikea-jobs-page/src/App.tsx
@@ -1,14 +1,9 @@
 import { useEffect } from 'react';
 import JobsList from './components/JobsList';
-import { AccessGate } from './components/AccessGate/AccessGate';
 
 function App() {
   useEffect(() => { window.scrollTo(0, 0); }, []);
-  return (
-    <AccessGate>
-      <JobsList />
-    </AccessGate>
-  );
+  return <JobsList />;
 }
 
 export default App;