+++ ikea-jobs-server/index.js
@@ -2,7 +2,6 @@ require('dotenv').config();
 const express = require('express');
 const cors = require('cors');
 const applicationRoutes = require('./src/routes/applicationRoutes');
-const accessRoutes = require('./src/routes/accessRoutes');
 
 const app = express();
 
@@ -10,7 +9,6 @@ app.use(cors());
 app.use(express.json());
 
 app.use('/api', applicationRoutes);
-app.use('/api', accessRoutes);
 
 const PORT = process.env.PORT || 3001;
 app.listen(PORT, () => {